test: identify services under test w/labels and setup conditionally (#486)
Modifies the tests to use labels to identify each service under
test. We then update the makefile to call ginkgo with a label filter,
and can then programmatically detect if (for example) the storage
service is under test. When the storage service is under test, we set
up the test resources. When the storage service is not under test, we
do not do that.

Previously we always stood up storage service test resources, which
caused errors in places where the storage service was not deployed
even if we weren't specifically testing the storage service.

+++ momento/test_helpers/shared_context.go
@@ -26,6 +26,7 @@ type SharedContext struct {
 	TopicClient                     momento.TopicClient
 	CacheName                       string
 	StoreName                       string
+	CleanupStore                    bool
 	Ctx                             context.Context
 	DefaultTtl                      time.Duration
 	Configuration                   config.Configuration
@@ -139,11 +140,12 @@ func (shared SharedContext) CreateDefaultCaches() {
 	}
 }
 
-func (shared SharedContext) CreateDefaultStores() {
+func (shared *SharedContext) CreateDefaultStores() {
 	_, err := shared.StorageClient.CreateStore(shared.Ctx, &momento.CreateStoreRequest{StoreName: shared.StoreName})
 	if err != nil {
 		panic(err)
 	}
+	shared.CleanupStore = true
 }
 
 func (shared SharedContext) Close() {
@@ -155,10 +157,12 @@ func (shared SharedContext) Close() {
 	if err != nil {
 		panic(err)
 	}
-	_, err = shared.StorageClient.DeleteStore(shared.Ctx, &momento.DeleteStoreRequest{StoreName: shared.StoreName})
-	if err != nil {
-		if err.(momento.MomentoError).Code() != momento.StoreNotFoundError {
-			panic(err)
+	if shared.CleanupStore {
+		_, err = shared.StorageClient.DeleteStore(shared.Ctx, &momento.DeleteStoreRequest{StoreName: shared.StoreName})
+		if err != nil {
+			if err.(momento.MomentoError).Code() != momento.StoreNotFoundError {
+				panic(err)
+			}
 		}
 	}
